requirements.txt
# --- Core Framework ---
Django==5.2.6
djangorestframework==3.17.1

# --- Configuration & Security ---
python-decouple==3.8
dj-database-url==3.0.1

# --- Database ---
psycopg2-binary==2.9.10

# --- Server / Deployment ---
gunicorn==23.0.0
whitenoise==6.11.0

# --- Django Dependencies (ติดตั้งอัตโนมัติมาพร้อม Django) ---
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2

python-dotenv
requests